Add AuthLink component, implement it on Nav, and fix Modal
The `AuthLink` component requires the following props: `to` and `content`, it then either renders a link to open the `AuthModal` if the user is not signed in, or the passed in `to` link that is passed in, the `to` link is not parsed if the user is signed in so you can pass in anything that the vanilla react-router `Link` component accepts. The component also accepts an optional `notice` boolean prop. If this prop is present the component will attach the notice "you must sign in before performing that action" will be attached to `location.state.notice`, we will have to actually render this on the modal if we want users to actually see it