Make major changes to directory structure
I think its really silly for the client package to be nested within the server package. This structure makes it hard to understand what is going on at a greater level within the application and leads to a false belief that the back-end server is somehow running the front-end server.

I restructured the directories such that the server and client are parallel in the packages directory. I believe this setup allows for a more intuitive understanding of how this project works.

- Simplified directory structure allows for other people to jump into this project and understand whats going on a little easier.
- Now you can install all the dependencies for both the client and the server through a single run of `yarn` rather than having to do `yarn` or `npm install` and then `yarn frontend-install` or `npm run frontend-install`.
- Scripts are significantly simplified in the root `package.json` because most of the more complex parameters are in the server and client's `package.json`
- All of the same scripts still work. `yarn dev` is still all you need to do to run the dev server

- You have to use yarn, sorry but it would be exceptionally complicated to allow for both to continue to be used here given the third party tools I've setup to help simplify the installation process. Almost all of the same commands work for yarn though and its great, I promise you'll learn to love it :sparkling_heart:
- The directory structure has changed and its going to take a little bit of getting used to.
- We're going to have to make some major changes to the `Dockerfile` in order to actually be able to deploy this. I'll gladly do this in the near future. I figure we don't have any urgent pushes that need to be made within the next 48 hours so the `Dockerfile` could probably wait a day or two so that y'all could check out the new structure and start getting used to it.

Directory structure will be very different after this is pulled in. If you hate this new structure I'd love to have a conversation about this. If you both hate it you can kill this PR and I promise I'll only cry about it for a couple days.